+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * New function align_test_files() to set time of signals in aligned re-recorded files
 * Parallel available on internal 'prep_X_bRlo_int()' function
+* Data frame are also returned by most functions
 
 # *baRulho 1.0.0*